Fix: use a revised macdeployqtfix from a fork
Which has a change that looks to be necessary to work with later
Python 3.x versions. The issue is that Python 3.x opens files
in a way that means strings are UTF-8 rather than bytes and
can not longer be directly read or wrote from/to files - thay
must be decoded and encoded first.

This should fix problems with packaging Mudlet MacOS builds.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>