Infrastructure: fix MacOS libzstd.1.dylib crashes (#111)
* Infrastructure: fix MacOS libzstd.1.dylib crashes

Copy libzip.5.dylib (and its dependencies: llibzstd.1.dylib and liblzma.5.dylib)
into a location where they seem to be needed/expected. For some reason
the existing tooling does not deal with the latter files. Also, "fix" the former
to look for those dependences in the same location in the Mudlet "Application
bundle" instead of referring into the "symlink-tree" used by the Homebrew
system which is not usable when the bundle is moved from the CI/CB
machine to an end-user's system without a Homebrew installation.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>