Fix: quote some script strings that have a space for PTBs
This was the final (I think) error in my recent MacOS CI build
failure fixes. I had removed some double quotation marks around some
a `bash` script variable forgetting that for PTB builds it contains
a space (between "Mudlet" and "PTB") in the file name and having that
then appear as two *words* made it seem like there was an extra
argument to the `install_name_tool` causing it to abort/fail the
build.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>